extract all of the handshake process to its own module
